Made non-integral nchan allowed (with 2014 likelihood). Energy estimator histograms still need to be for integral values of nchan, but events can have fractional nchan and background distributions can include bins for fractional nchan.
	modified:   include/nulike.h
	modified:   programs/nulike_test.f
	modified:   src/bginit.f
	modified:   src/bgspec.f
	modified:   src/bounds.f
	modified:   src/edisp.f
	modified:   src/edispinit.f
	modified:   src/init.f
	modified:   src/speclike.f

+++ src/edispinit.f
@@ -56,7 +56,7 @@
       hist_prob(:,:,analysis) = 0.d0
       do k = 1, nnchan_total(analysis)
         do i = 1, nbins_Ein
-          hist_nchan(i,k,analysis) = k - 1 + nchan_min(analysis)
+          hist_nchan(i,k,analysis) = k - 1 + nint(nchan_min(analysis))
           do j = 1, nbins_nchan(i)
             if (hist_nchan_temp(i,j) .eq. hist_nchan(i,k,analysis)) then
               hist_prob(i,k,analysis) = hist_prob_temp(i,j)
